+++ Asset Management Platform/Trade.cs
@@ -12,13 +12,13 @@ namespace Asset_Management_Platform
         public string BuyOrSell;
         public Security Security;
         public string Ticker;
-        public int Shares;
+        public decimal Shares;
         public string Terms;
         public decimal Limit;
         public string OrderDuration;
 
 
-        public Trade(string buyOrSell, Security securityToAdd, string ticker, int shares, string terms, decimal limit, string orderDuration)
+        public Trade(string buyOrSell, Security securityToAdd, string ticker, decimal shares, string terms, decimal limit, string orderDuration)
         {
             BuyOrSell = buyOrSell;
             Security = securityToAdd;
